backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.3.0
# pgvector==0.2.4  # Optional - not needed for current implementation

# Pydantic for schemas
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
anthropic==0.18.0

# Content Extraction
beautifulsoup4==4.12.3
lxml  # Version managed by trafilatura
pypdf2==3.0.1
pdfplumber==0.10.3
requests==2.31.0
aiohttp==3.9.1

# Twitter (using official API v2)
tweepy==4.14.0

# arXiv
arxiv==2.1.0

# Web scraping
trafilatura==1.6.3
readability-lxml==0.8.1

# Utilities
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2
httpx==0.26.0
aiofiles==23.2.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Rate limiting
slowapi==0.1.9
